commit25-10-22

+++ app/views.py
@@ -13,16 +13,19 @@ from django.contrib.auth.decorators import login_required
 from django.utils.decorators import method_decorator
 User = get_user_model()
 
+votp = None
 # Home
 
 
 def home(request):
     jeans = Product.objects.filter(category="jeans")
-    return render(request, 'app/home.html', {"jeans": jeans})
+    mobile = Product.objects.filter(category="mobile")
+    return render(request, 'app/home.html', {"jeans": jeans, "mobile":mobile})
 
 
 def mobile(request):
-    return render(request, "mobile.html")
+    mobile = Product.objects.filter(category="mobile")
+    return render(request, "app/mobile.html", {"mobile":mobile})
 # ------------------------------------------------------------------------------
 
 
@@ -37,49 +40,59 @@ def send_otp(phone_no):
     return otp
 
 
+
+
 def customerregistration(request):
     # Registration form
-    if request.method == "POST":
-        global forma
-        forma = UserRegisterForm(request.POST)
-        if forma.is_valid():
-            phone_number = forma.cleaned_data['mobile_no']
-            global otp
-            otp = send_otp(phone_number)
-            print(otp)
-            messages.success(request, 'Otp has been sent to your mobile n.o')
+    if request.user.is_anonymous:
+        if request.method == "POST":
+            global forma
+            forma = UserRegisterForm(request.POST)
+            if forma.is_valid():
+                phone_number = forma.cleaned_data['mobile_no']
+                global otp
+                otp = send_otp(phone_number)
+                print(otp)
+                messages.success(request, 'Otp has been sent to your mobile n.o')
+                global votp
+                votp = random.randint(1000,10000)
+                return redirect("/verify_otp/"+str(votp))
+        else:
+            forma = UserRegisterForm()
 
-            return redirect(verify_otp)
+        return render(request, 'app/customerregistration.html', {"forms": forma})
     else:
-        forma = UserRegisterForm()
-
-    return render(request, 'app/customerregistration.html', {"forms": forma})
+        return redirect("/")
 
 
-def verify_otp(request):
+def verify_otp(request,pk):
     # verify otp and save user to database
-    print(request)
-    print(request.data)
+    if votp is not None:
+        if int(pk) == int(votp):
+            if request.method == 'POST':
+                form = VerifyOtpForm(request.POST)
+                if form.is_valid():
+                    otp_value = form.cleaned_data['Otp']
+                    if int(otp_value) == otp:
 
-    if request.method == 'POST':
-        form = VerifyOtpForm(request.POST)
-        if form.is_valid():
-            otp_value = form.cleaned_data['Otp']
-            if int(otp_value) == otp:
+                        forma.save()
+                        messages.success(request, 'You are registered successfully')
+                        return redirect("/login/")
+                    else:
+                        messages.warning(request, "Sorry invalid Otp please try again")
+                        return redirect(customerregistration)
 
-                forma.save()
-                messages.success(request, 'You are registered successfully')
-                return redirect("/login/")
             else:
-                messages.error(request, "Sorry invalid Otp please try again")
-                return redirect(customerregistration)
-
+                form = VerifyOtpForm()
+                return render(request, 'app/verify_otp.html', {'form': form})
+        else:
+            return redirect(customerregistration)
     else:
-        form = VerifyOtpForm()
-        return render(request, 'app/verify_otp.html', {'form': form})
+            return redirect(customerregistration)
 
+   
 
-@method_decorator(login_required(login_url='/'), name='dispatch')
+@method_decorator(login_required(login_url='login'), name="dispatch")
 class ProfileUpdate(UpdateView):
     # User can edit his current profile
     model = User
@@ -88,42 +101,55 @@ class ProfileUpdate(UpdateView):
     success_url = '/'
     extra_context = {"active": "btn-primary"}
 
+    def dispatch(self, request, *args, **kwargs):
+        obj =  self.get_object()
+        if obj != self.request.user:
+            return redirect("/")
+        else:
+            return super().dispatch(request, *args, *kwargs)
+
 # ------------------------------------------------------------------------------
 
 # Addresses of User.
 
 
-@login_required(login_url='login')
+
 def add_address(request, pk):
     # Add address
-    if request.user.id == pk:
-        if request.method == "POST":
-            form = AddOtherAddressForm(request.POST)
-            if form.is_valid():
-                obj = form.save(commit=False)
-                obj.user_id = pk
-                obj.save()
-                return redirect('profile', pk=pk)
+    if not request.user.is_anonymous :
+        if request.user.id == pk:
+            if request.method == "POST":
+                form = AddOtherAddressForm(request.POST)
+                if form.is_valid():
+                    obj = form.save(commit=False)
+                    obj.user_id = pk
+                    obj.save()
+                    return redirect('profile', pk=pk)
+            else:
+                form = AddOtherAddressForm()
+            return render(request, 'app/address.html', {'form': form})
         else:
-            form = AddOtherAddressForm()
-        return render(request, 'app/address.html', {'form': form})
+            return redirect('/')
     else:
-        messages.warning(request, "don't try to login in others account")
+        messages.warning(request, "Don't try to login in others account")
         return redirect('/login')
 
 
-@login_required(login_url='login')
+
 def manage_address(request, pk):
     # Show and option for delete or edit created address
-    if request.user.id == pk:
-        add = Customer.objects.filter(user_id=pk)
-        return render(request, "app/addressmanage.html", {'active': 'btn-primary', "add": add})
+    if not request.user.is_anonymous :
+        if request.user.id == pk:
+            add = Customer.objects.filter(user_id=pk)
+            return render(request, "app/addressmanage.html", {'active': 'btn-primary', "add": add})
+        else:
+            return redirect('/')
     else:
-        messages.warning(request, "don't try to login in others account")
-        return redirect('/login')
+            messages.warning(request, "Don't try to login in others account")
+            return redirect('/login')
 
 
-@method_decorator(login_required(login_url='/'), name='dispatch')
+@method_decorator(login_required(login_url='login'), name='dispatch')
 class AddressUpdate(UpdateView):
     # Edit address
     model = Customer
@@ -133,14 +159,28 @@ class AddressUpdate(UpdateView):
     cid = Customer.id
     extra_context = {"active": "btn-primary", "cid": cid}
 
+    def dispatch(self, request, *args, **kwargs):
+        obj =  self.get_object()
+        if obj.user != self.request.user:
+            return redirect("/")
+        else:
+            return super().dispatch(request, *args, *kwargs)
 
-@method_decorator(login_required(login_url='/'), name='dispatch')
+
+@method_decorator(login_required(login_url='login'), name='dispatch')
 class AddressDelete(DeleteView):
     # Delte Address
     model = Customer
     context_object_name = 'object'
     template_name = 'app/addressdelete.html'
     success_url = '/'
+
+    def dispatch(self, request, *args, **kwargs):
+        obj =  self.get_object()
+        if obj.user != self.request.user:
+            return redirect("/")
+        else:
+            return super().dispatch(request, *args, *kwargs)
 
 
 # ------------------------------------------------------------------------------
@@ -159,20 +199,23 @@ def product_detail(request, pk):
         return render(request, 'app/productdetail.html', {"product": product})
 
 
-# @login_required(login_url='login')
 def add_to_cart(request, pk):
-    product = Product.objects.get(id=pk)
-    user = request.user
-    item = Cart.objects.filter(
-        Q(user=request.user) & Q(product=product.id)).exists()
-    if item:
-        return redirect('/showcart')
+    if not request.user.is_anonymous :
+            product = Product.objects.get(id=pk)
+            user = request.user
+            item = Cart.objects.filter(
+                Q(user=request.user) & Q(product=product.id)).exists()
+            if item:
+                return redirect('/showcart')
+            else:
+                item = Cart(user=user, product=product).save()
+                return redirect('/showcart')
     else:
-        item = Cart(user=user, product=product).save()
-        return redirect('/showcart')
+            messages.warning(request, "Don't try to login in others account")
+            return redirect('/login')
 
 
-# @login_required(login_url='login')
+@login_required(login_url='login')
 def showcart(request):
     if request.user.is_authenticated:
         user = request.user
@@ -193,76 +236,92 @@ def showcart(request):
         return render(request, "app/emptycart.html")
 
 # Cart Quantity
-# @login_required( login_url='login')
-
 
 def plus(request):
-    if request.method == 'GET':
-        prod_id = request.GET['prod_id']
-        c = Cart.objects.get(Q(product=prod_id) & Q(user=request.user))
-        c.quantity += 1
-        c.save()
-        cart_product = [p for p in Cart.objects.all() if p.user ==
-                        request.user]
-    amount = 0
-    shipping_amount = 70
-    for p in cart_product:
-        amount += p.quantity*p.product.discounted_price
-   
-    data = {
-        'quantity': c.quantity,
-        'amount': amount,
-        'totalamount': amount+shipping_amount
-    }
-    return JsonResponse(data)
+    if not request.user.is_anonymous :
+            if request.method == 'GET':
+                try:
+                    prod_id = request.GET['prod_id']
+                    c = Cart.objects.get(Q(product=prod_id) & Q(user=request.user))
+                    c.quantity += 1
+                    c.save()
+                    cart_product = [p for p in Cart.objects.all() if p.user ==
+                                    request.user]
+                except:
+                    return redirect("/")
+                amount = 0
+                shipping_amount = 70
+                for p in cart_product:
+                    amount += p.quantity*p.product.discounted_price
+            
+                data = {
+                    'quantity': c.quantity,
+                    'amount': amount,
+                    'totalamount': amount+shipping_amount
+                }
+                return JsonResponse(data)
+            
 
-# @login_required( login_url='login')
-
+    else:
+        messages.warning(request, "Don't try to login in others account")
+        return redirect('/login')
 
 def minus(request):
-    if request.method == 'GET':
-        prod_id = request.GET['prod_id']
-        c = Cart.objects.get(Q(product=prod_id) & Q(user=request.user))
-        c.quantity -= 1
-        c.save()
-        cart_product = [p for p in Cart.objects.all() if p.user ==
-                        request.user]
-    amount = 0
-    shipping_amount = 70
-    for p in cart_product:
-        amount += p.quantity*p.product.discounted_price
-
-    data = {
-        'quantity': c.quantity,
-        'amount': amount,
-        'totalamount': amount+shipping_amount
-    }
-    return JsonResponse(data)
-
-# @login_required( login_url='login')
-
-
-def remove_cart(request):
-    if request.method == 'GET':
-        prod_id = request.GET['prod_id']
-        c = Cart.objects.get(Q(product=prod_id) & Q(user=request.user))
-        c.delete()
-        amount = 0.0
-        shipping_amount = 70.0
-        cart_product = [p for p in Cart.objects.all() if p.user ==
-                        request.user]
+    if not request.user.is_anonymous :
+        if request.method == 'GET':
+            try:
+                prod_id = request.GET['prod_id']
+                c = Cart.objects.get(Q(product=prod_id) & Q(user=request.user))
+                c.quantity -= 1
+                c.save()
+                cart_product = [p for p in Cart.objects.all() if p.user ==
+                                request.user]
+            except:
+                return redirect("/")
+        amount = 0
+        shipping_amount = 70
         for p in cart_product:
-            tempamount = (p.quantity * p.product.discounted_price)
-            amount += tempamount
+            amount += p.quantity*p.product.discounted_price
 
         data = {
+            'quantity': c.quantity,
             'amount': amount,
             'totalamount': amount+shipping_amount
         }
-
         return JsonResponse(data)
     else:
-        return HttpResponse("")
+        messages.warning(request, "Don't try to login in others account")
+        return redirect('/login')
+
+def remove_cart(request):
+    if not request.user.is_anonymous :
+        if request.method == 'GET':
+            try:
+                prod_id = request.GET['prod_id']
+                c = Cart.objects.get(Q(product=prod_id) & Q(user=request.user))
+                c.delete()
+                amount = 0.0
+                shipping_amount = 70.0
+                cart_product = [p for p in Cart.objects.all() if p.user ==
+                                request.user]
+            except:
+                return redirect("/")
+            for p in cart_product:
+                tempamount = (p.quantity * p.product.discounted_price)
+                amount += tempamount
+
+            data = {
+                'amount': amount,
+                'totalamount': amount+shipping_amount
+            }
+
+            return JsonResponse(data)
+        else:
+            return HttpResponse("")
+    else:
+        messages.warning(request, "Don't try to login in others account")
+        return redirect('/login')
+
 
 # Placed Order
 
@@ -281,25 +340,29 @@ def checkout(request):
     address = Customer.objects.filter(user=user)
     return render(request, 'app/checkout.html', {"carts": carts, "totalamount": total_amount, "address": address})
 
-
-@login_required(login_url='login')
 def paymentdone(request):
-    custid = request.GET.get("custid")
-    print(custid)
-    custom = Customer.objects.get(id=custid)
-    print(custom)
-    user = request.user
-    cart = Cart.objects.filter(user=user)
-    for p in cart:
-        Product_Status(user=user, product=p.product,
-                       quantity=p.quantity, customer=custom).save()
-        p.delete()
-    return redirect('/orders')
+    if not request.user.is_anonymous :
+        try:
+            custid = request.GET.get("custid")
+            custom = Customer.objects.get(id=custid)
+            print(custom)
+            user = request.user
+            cart = Cart.objects.filter(user=user)
+        except:
+             return redirect('/')
+        for p in cart:
+            Product_Status(user=user, product=p.product,
+                        quantity=p.quantity, customer=custom).save()
+            p.delete()
+        return redirect('/orders')
+        
 
+    else:
+        messages.warning(request, "Don't try to login in others account")
+        return redirect('/login')
 
 def buy_now(request, pk):
     pass
-
 
 @login_required(login_url='login')
 def orders(request):
